+++ src/main/kotlin/no/iktdev/kammich/models/shared/RemoteFile.kt
@@ -1,8 +1,8 @@
 package no.iktdev.kammich.models.shared
 
 data class RemoteFile(
-    val id: Int,
-    val deviceId: Int,
+    val id: Long,
+    val deviceId: Long,
     val fileName: String,
     val uploaded: Boolean = false
 ) {
